Implement Laplace approx for classification, bug fixes
For the Laplace approximation, most everything is done, EXCEPT:
I need to implement derivatives of the negative log marginal likelihood
with respect to the likelihood function hyperparameters, and for
marginal effects, we need to deal with derivatives of the likelihood
(NOT the log likelihood) with respect to f.
On the NLML, I haven't implemented this currently since neither the
logistic nor the probit likelihood have hyperparameters; however,
I'll need to put it in at some point since ostensibly we could need to
deal with a likelihood that does have hypers.
On the likelihood derivatives, currently I just have the derivative for
the logistic likelihood hard-coded in the marginal effects method,
since that's the only binary likelihood currently available in the
package. But, eventually we'll need to farm this out to the
LikelihoodFunction classes.

I fixed a small bug in the CovSEard cov() method that could occur if X
has only one predictor: The M matrix could end up being 0x0, so now we
ensure that doesn't happen.

I also changed whether the GPModel constructor or the child class
constructor is responsible for calling train(); now we rely on the
child classes to do this, to allow them to make additional checks after
the outcomes and design matrix have been constructed.