Ensure MeanLinear returns vector, not 1 col matrix
Theoretically a vector and a one column matrix are the same, but not in
terms of how they are treated by some functions in R. In places where
a vector and one-column matrix would be treated differently by R
functions, we want it treated like a vector (e.g. diag(mu)).